CameraController has 'Minecraft' controls

+++ ZPGine/src/Core/Window.h
@@ -36,7 +36,7 @@ public:
     virtual bool IsVSync() const = 0;
 
     virtual void* GetNativeWindow() const = 0;
-    virtual f32 GetAspectRatio() const { return GetWidth() / (f32)GetHeight(); }
+    virtual f32 GetAspectRatio() const { return static_cast<f32>(GetWidth()) / static_cast<f32>(GetHeight()); }
 
     static Scope<Window> Create(const WindowProps& props = WindowProps());
 };
